client interval: add entry
Fixed templates. So templates are really only accessed if vulnerable default config options are found. In which case, templates are mainly there for line numbers and the red error line. Given this, there are 2 situations templates are accessed. The first is when defaults are explicitly set to a vulnerable state (in which case the template would provide a match and a line number/red line will be given). The third is when defaults are implicitly set, in which case no template would be matched since the config option isn't in the file. For some reason I didn't handle this properly before so now sve simply outputs the config option (not its value, just the keyword) in the red error line and no line number is given.

Default red error prompt is also changed from 'missing:' to 'implicit:'.